+++ client/src/images/sprites/loading-loop-300.png
