+++ client/src/images/sprites/network-status-60.png
